refactor(output-view): 출력만을 위한 메소드는 OutputView 로 분리
