+++ src/main/java/view/InputView.java
@@ -1,6 +1,5 @@
 package view;
 
-import dto.LottoDto;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Scanner;
@@ -9,13 +8,12 @@ import java.util.stream.Collectors;
 public class InputView {
     private static final Scanner SCANNER = new Scanner(System.in);
 
-    private static final String ERROR_MESSAGE_PREFIX = "[ERROR] ";
+    private static final String SPLIT_DELIMITER = ", ";
+
     private static final String ERROR_MESSAGE_FOR_INVALID_NUMBER = "숫자만 입력해주세요.";
     private static final String MESSAGE_TO_GET_INPUT_MONEY = "구입금액을 입력해 주세요.";
-    private static final String MESSAGE_FOR_LOTTO_COUNT = "%d개를 구매했습니다.%n";
     private static final String MESSAGE_FOR_WINNING_LOTTO_NUMBERS = "지난 주 당첨 번호를 입력해 주세요.";
     private static final String MESSAGE_FOR_BONUS_NUMBER = "보너스 볼을 입력해 주세요.";
-    private static final String SPLIT_DELIMITER = ", ";
 
     public static int scanInputMoney() {
         System.out.println(MESSAGE_TO_GET_INPUT_MONEY);
@@ -27,19 +25,10 @@ public class InputView {
         }
     }
 
-    public static void printException(Exception exception) {
-        System.out.println(ERROR_MESSAGE_PREFIX + exception.getMessage());
-    }
-
-    public static void printLottoQuantity(int count) {
-        System.out.printf(MESSAGE_FOR_LOTTO_COUNT, count);
-    }
-
     public static List<Integer> scanWinningNumbers() {
-        System.out.print(System.lineSeparator());
-        System.out.println(MESSAGE_FOR_WINNING_LOTTO_NUMBERS);
-        String userInput = SCANNER.nextLine();
+        System.out.println(System.lineSeparator() + MESSAGE_FOR_WINNING_LOTTO_NUMBERS);
 
+        String userInput = SCANNER.nextLine();
         try {
             return Arrays.stream(userInput.split(SPLIT_DELIMITER))
                     .map(Integer::parseInt)
@@ -57,21 +46,5 @@ public class InputView {
         } catch (NumberFormatException exception) {
             throw new IllegalArgumentException(ERROR_MESSAGE_FOR_INVALID_NUMBER);
         }
-    }
-
-    public static void printLottos(List<LottoDto> lottoDtos) {
-        for (LottoDto lottoDto : lottoDtos) {
-            printSingleLotto(lottoDto);
-        }
-    }
-
-    public static void printSingleLotto(LottoDto lottoDto) {
-        System.out.println("[" + joinNumbers(lottoDto.getLottoNumbers()) + "]");
-    }
-
-    private static String joinNumbers(List<Integer> numbers) {
-        return numbers.stream()
-                .map(String::valueOf)
-                .collect(Collectors.joining(SPLIT_DELIMITER));
     }
 }
